+++ packages/package-is-installable/src/index.ts
@@ -3,16 +3,13 @@ import {
   skippedOptionalDependencyLogger,
 } from '@pnpm/core-loggers'
 import { PackageManifest } from '@pnpm/types'
-import { splitNodeId } from '@pnpm/utils'
 import installChecks = require('pnpm-install-checks')
-import { ResolvedPackagesByPackageId } from './resolveDependencies'
 
 export default async function getIsInstallable (
   pkgId: string,
   pkg: PackageManifest,
   options: {
     engineStrict: boolean,
-    resolvedPackagesByPackageId: ResolvedPackagesByPackageId,
     nodeId: string,
     nodeVersion: string,
     optional: boolean,
@@ -45,7 +42,7 @@ export default async function getIsInstallable (
         name: pkg.name,
         version: pkg.version,
       },
-      parents: nodeIdToParents(options.nodeId, options.resolvedPackagesByPackageId),
+      parents: undefined, // TODO: remove this field from the type
       prefix: options.prefix,
       reason: warn.code === 'ENOTSUP' ? 'unsupported_engine' : 'unsupported_platform',
     })
@@ -56,20 +53,4 @@ export default async function getIsInstallable (
   if (options.engineStrict) throw warn
 
   return true
-}
-
-export function nodeIdToParents (
-  nodeId: string,
-  resolvedPackagesByPackageId: ResolvedPackagesByPackageId,
-) {
-  const pkgIds = splitNodeId(nodeId).slice(2, -2)
-  return pkgIds
-    .map((pkgId) => {
-      const pkg = resolvedPackagesByPackageId[pkgId]
-      return {
-        id: pkg.id,
-        name: pkg.name,
-        version: pkg.version,
-      }
-    })
 }
