+++ packages/package-is-installable/tslint.json
@@ -1,0 +1,3 @@
+{
+  "extends": "@pnpm/tslint-config"
+}
