+++ packages/package-is-installable/typings/local.d.ts
@@ -1,0 +1,4 @@
+declare module 'pnpm-install-checks' {
+  const anything: any;
+  export = anything;
+}
